Start splitting commands into modules.
