+++ meta/embeds.js
@@ -289,21 +289,3 @@ exports.command.feedback = message => {
 		}
 	})
 }
-
-exports.command.ping = message => {
-	return({
-		title: `${pfx}ping`,
-		description: `Ping the bot.`,
-		url: "https://allydiscord.github.io/docs/commands/utility/ping/",
-		color: 0x0099FF,
-		thumbnail: {
-			url: "https://imgur.com/YVRMcUD.png"
-		},
-		fields: [],
-		timestamp: new Date(),
-		footer: {
-			text: `Requested by ${message.author.username} | ${nme} v${vrs}`,
-			icon_url: message.author.avatarURL()
-		}
-	})
-}
